even more testing, almost done

+++ service/answer_service.py
@@ -18,6 +18,8 @@ async def answer_question(request : AnswerRequest)-> Optional[int]:
         if user_answered_check:
             raise ValueError ("User already answered question, use update answer instead")
         else:
+            if request.answer_id > 4 or request.answer_id < 0:
+                raise ValueError('answer_id must be between 1 and 4')
             return await poll_client.answer_question(request)
     else:
         raise ValueError ("User needs to be registered")
@@ -28,13 +30,18 @@ async def update_answer(request : AnswerRequest)->Optional[int]:
     if user.is_registered:
         user_answered_check = await poll_client.check_user_answered(user.id, request.question_id)
         if user_answered_check:
-            return await poll_client.update_answer(request.question_id, request.answer_id, request.user_id)
+            if request.answer_id > 4 or request.answer_id < 0:
+                return await poll_client.update_answer(request.question_id, request.answer_id, request.user_id)
+            raise ValueError('answer_id must be between 1 and 4')
         else:
             raise ValueError ("User hasn't answered this question yet, use answer question instead")
     else:
         raise ValueError ("User needs to be registered")
 
 async def delete_answer(user_id:int, question_id:int)->Optional[int]:
+    user : User = await user_repository.get_user_by_id(user_id)
+    if not user:
+        raise ValueError("User not found")
     user_answered_check = await poll_client.check_user_answered(user_id, question_id)
     if user_answered_check:
         return await poll_client.delete_answer(user_id,question_id)
